+++ packages/wallets/ledger-hid/src/icons/index.ts
@@ -1,0 +1,2 @@
+export { default as WalletIcon } from './ledger.svg';
+export { default as WalletIconInverted } from './ledger-inverted.svg';