+++ packages/web3-react/src/hooks/useConnectorInfo.ts
@@ -2,10 +2,7 @@ import { useAccount } from 'wagmi';
 import { SafeAppConnector } from '@gnosis.pm/safe-apps-web3-react';
 import { InjectedConnector } from '@web3-react/injected-connector';
 import { WalletLinkConnector } from '@web3-react/walletlink-connector';
-import {
-  LedgerHQConnector,
-  LedgerHQFrameConnector,
-} from '@reef-knot/ledger-connector';
+import { LedgerHQFrameConnector } from '@reef-knot/ledger-connector';
 import { useWeb3 } from './useWeb3';
 import { PROVIDER_NAMES } from '../constants';
 import {
@@ -22,7 +19,6 @@ type ConnectorInfo = {
   providerName?: string;
   isConnectedViaWagmi: boolean;
   isGnosis: boolean;
-  isLedger: boolean;
   isLedgerLive: boolean;
   isWalletLink: boolean;
   isCoinbase: boolean;
@@ -46,7 +42,6 @@ export const useConnectorInfo = (): ConnectorInfo => {
   // === WEB3-REACT connectors BEGIN
   const isGnosis = active && connector instanceof SafeAppConnector;
   const isLedgerLive = active && connector instanceof LedgerHQFrameConnector;
-  const isLedger = connector instanceof LedgerHQConnector;
 
   // WalletLink is used by Coinbase, but it can be used by other wallets too.
   const isWalletLink = active && connector instanceof WalletLinkConnector;
@@ -78,7 +73,6 @@ export const useConnectorInfo = (): ConnectorInfo => {
     if (isDappBrowser) return undefined;
 
     if (isGnosis) return PROVIDER_NAMES.GNOSIS;
-    if (isLedger) return PROVIDER_NAMES.LEDGER;
     if (isLedgerLive) return PROVIDER_NAMES.LEDGER_HQ_LIVE;
     if (isImToken) return PROVIDER_NAMES.IM_TOKEN;
     if (isTrust) return PROVIDER_NAMES.TRUST;
@@ -107,7 +101,6 @@ export const useConnectorInfo = (): ConnectorInfo => {
     isConnectedViaWagmi,
 
     isGnosis,
-    isLedger,
     isLedgerLive,
     isWalletLink,
     isCoinbase,