+++ packages/web3-react/test/hooks/useConnectorInfo.test.ts
@@ -3,9 +3,8 @@ jest.mock('wagmi');
 
 import { renderHook } from '@testing-library/react-hooks';
 import { SafeAppConnector } from '@gnosis.pm/safe-apps-web3-react';
-import { WalletLinkConnector } from '@web3-react/walletlink-connector';
 import { InjectedConnector } from '@web3-react/injected-connector';
-import { LedgerHQConnector, LedgerHQFrameConnector } from '@reef-knot/ledger-connector';
+import { LedgerHQFrameConnector } from '@reef-knot/ledger-connector';
 import { useConnectorInfo } from '../../src/hooks/useConnectorInfo';
 import { useWeb3 } from '../../src/hooks/useWeb3';
 import { useAccount } from 'wagmi';
@@ -44,33 +43,12 @@ describe('useConnectorInfo', () => {
   });
 
   test('should detect ledger live', async () => {
-    mockConnector(LedgerHQConnector);
-    const { result } = renderHook(() => useConnectorInfo());
-    const { providerName, isLedger, isConnectedViaWagmi, ...rest } = result.current;
-
-    expect(isLedger).toBe(true);
-    expect(Object.values(rest).includes(true)).toBeFalsy();
-  });
-
-  test('should detect ledger live', async () => {
     mockConnector(LedgerHQFrameConnector);
     const { result } = renderHook(() => useConnectorInfo());
     const { providerName, isLedgerLive, isConnectedViaWagmi, ...rest } =
       result.current;
 
     expect(isLedgerLive).toBe(true);
-    expect(Object.values(rest).includes(true)).toBeFalsy();
-  });
-
-  test('should detect coinbase', async () => {
-    mockConnector(WalletLinkConnector);
-    window.ethereum = { isCoinbaseWallet: true };
-
-    const { result } = renderHook(() => useConnectorInfo());
-    const { providerName, isCoinbase, isWalletLink, isConnectedViaWagmi, ...rest } =
-      result.current;
-
-    expect(isCoinbase).toBe(true);
     expect(Object.values(rest).includes(true)).toBeFalsy();
   });
 
@@ -85,6 +63,7 @@ describe('useConnectorInfo', () => {
 
   test('should detect MetaMask', async () => {
     mockConnector(InjectedConnector);
+    // @ts-expect-error
     window.ethereum = { isMetaMask: true };
 
     const { result } = renderHook(() => useConnectorInfo());
@@ -93,46 +72,6 @@ describe('useConnectorInfo', () => {
 
     expect(isInjected).toBe(true);
     expect(isMetamask).toBe(true);
-    expect(Object.values(rest).includes(true)).toBeFalsy();
-  });
-
-  test('should detect imToken', async () => {
-    mockConnector(InjectedConnector);
-    window.ethereum = { isImToken: true };
-
-    const { result } = renderHook(() => useConnectorInfo());
-    const { providerName, isInjected, isImToken, isConnectedViaWagmi, ...rest } =
-      result.current;
-
-    expect(isInjected).toBe(true);
-    expect(isImToken).toBe(true);
-    expect(Object.values(rest).includes(true)).toBeFalsy();
-  });
-
-  test('should detect Trust', async () => {
-    mockConnector(InjectedConnector);
-    window.ethereum = { isTrust: true };
-
-    const { result } = renderHook(() => useConnectorInfo());
-    const { providerName, isInjected, isTrust, isConnectedViaWagmi, ...rest } =
-      result.current;
-
-    expect(isInjected).toBe(true);
-    expect(isTrust).toBe(true);
-    expect(Object.values(rest).includes(true)).toBeFalsy();
-  });
-
-  test('should detect XDEFI', async () => {
-    mockConnector(InjectedConnector);
-    window.ethereum = { isXDEFI: true };
-    window.xfi = {};
-
-    const { result } = renderHook(() => useConnectorInfo());
-    const { providerName, isInjected, isXdefi, isConnectedViaWagmi, ...rest } =
-      result.current;
-
-    expect(isInjected).toBe(true);
-    expect(isXdefi).toBe(true);
     expect(Object.values(rest).includes(true)).toBeFalsy();
   });
 
@@ -149,6 +88,7 @@ describe('useConnectorInfo', () => {
   test('should detect mobile dapp browser', async () => {
     mockIsMobileOrTablet.mockReturnValue(true);
     mockConnector(InjectedConnector);
+    // @ts-expect-error
     window.ethereum = {};
 
     const { result } = renderHook(() => useConnectorInfo());