+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @reef-knot/types
 
+## 1.4.0
+
+### Minor Changes
+
+- Use Ledger hid wallet adapter, based on wagmi
+
 ## 1.3.0
 
 ### Minor Changes